move object factory specs to all be in one place
